Fix double-close when using File()
Using os.NewFile() can close the underlaying fd used to create it in
several situations. The fd used is the same we try to close on our
Close() method, accounting for double close on the fd: once when the
os.File is GC/closed, another when our Close() method is called.

This gets worse if the user for some reason uses our File() method more
than once: all os.File() use the same underlying fd and closing one will
invalidate the others.

To fix this, this patch just uses dup() on the fd before passing it to
os.NewFile(). This way, each instance can (and must) call Close on the
os.File and also our Close method when dealing with the event is over.

A simple way to reproduce this issue before this patch, is applying the
following diff the the main.go example:

	diff --git a/cmd/main.go b/cmd/main.go
	index 9f7dfb2..6b85bcd 100644
	--- a/cmd/main.go
	+++ b/cmd/main.go
	@@ -54,13 +54,22 @@ func main() {
	                        return "", nil
	                }

	-               defer data.Close()
	+               dataFile := data.File()
	+               //defer data.Close()

	                path, err := data.GetPath()
	                if err != nil {
	                        return "", err
	                }

	+               if err := data.Close(); err != nil {
	+                       return "", fmt.Errorf("Err data.Close(): %v", err)
	+               }
	+
	+               if err := dataFile.Close(); err != nil {
	+                       return "", fmt.Errorf("Err 1 dataFile.Close(): %v", err)
	+               }
	+
	                if data.MatchMask(unix.FAN_CLOSE_WRITE) || data.MatchMask(unix.FAN_MODIFY) {
	                        return fmt.Sprintf("PID:%d %s", data.GetPID(), path), nil
	                }

Without this patch, the program fails with:
	error: Err 1 dataFile.Close(): close : bad file descriptor

After this patch, it works as expected.

Fixes: #5

Signed-off-by: Rodrigo Campos <rodrigo@kinvolk.io>